// frontend/src/lib/api/types/listAlertEventsParams.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Momentum Dashboard API
 * Contract-first spec. Clients must be generated from this file. All endpoints are local-only and authenticated as N/A for now.
 * OpenAPI spec version: 0.8.0
 */
import type { ListAlertEventsSeverity } from './listAlertEventsSeverity';
import type { ListAlertEventsDigestBucket } from './listAlertEventsDigestBucket';
import type { ListAlertEventsMode } from './listAlertEventsMode';
import type { ListAlertEventsChannelStatus } from './listAlertEventsChannelStatus';

export type ListAlertEventsParams = {
symbol?: string;
trading_date?: string;
trading_date_from?: string;
trading_date_to?: string;
severity?: ListAlertEventsSeverity;
digest_bucket?: ListAlertEventsDigestBucket;
mode?: ListAlertEventsMode;
rule_code?: string;
next_action_code?: string;
channel_status?: ListAlertEventsChannelStatus;
profile?: string;
limit?: number;
cursor?: string;
};